Update hyperparameters and adjust range in generate_S15G3W1_scaled.py script
The 'N_SAMPLE' in FIXED_HP section has been reduced from 64 to 16. This change comes along with the modification of the 'reg_fn_alpha' range from (-5, 10) to (-3, 3). Adjustments are also made to 'reg_proj', 'seed', 'normalization_fn', and 'normalization_nu_fn' properties in VARIANLE_HP. These modifications aim to improve the efficiency and accuracy of the model training process.