Add new hyperparameter generation script for S15G3W1_LogEPathLen
This commit adds a new script to generate hyperparameters for the model "S15G3W1_LogEPathLen". This script has two main sections, FIXED_HP and VARIABLE_HP, defining fixed and variable hyperparameters respectively. For each combination of variable hyperparameters, the script generates a unique hash, which is then used to store a pickle file containing that particular set of hyperparameters.