Refactor functions to models in nu_rescale module
Various functions in the 'nu_rescale.py' script have been refactored into subclasses of the tf.keras.Model class. This restructuring improves code organization, promotes greater consistency with the rest of the TensorFlow eco-system, and opens the way for the potential use of these models in workflows that require TensorFlow model objects. The callback printing of 'alpha_range' in 'callbacks.py' and additional HP methods in 'losses.py' are also updated for better tracking of model's parameters during the training session.