Fixed method return and updated parameters in callbacks
Defect in "losses.py" was fixed by adding a return statement in the update function. This change is also reflected in "callbacks.py" script, where the 'reg_fn_alpha' and 'reg_fn_beta' parameters are replaced by a call to the updated method, streamlining the updating of results with more straight-forward parameter retrieval during training sessions.